+++ Beedraz/semantics_II/trunk/src/test/java/org/beedra_II/property/number/integer/long64/StubEditableLongBeed.java
@@ -26,7 +26,7 @@ public class StubEditableLongBeed extends EditableLongBeed {
   }
 
   @Override
-  public boolean isAcceptable(Integer goal) {
+  public boolean isAcceptable(Long goal) {
     return goal != null && goal > 0;
   }
